Move quickstart functionality to disposable
* and out of workflow
